[vmbase] Adjust refill layout size in memory allocation
This cl adjusts the refill layout size to the next power of two
when allocating memory. This is necessary because the actual size
of the memory allocated in the subsequent call `alloc_aligned` is
the power of two.
Without this adjustment, the `alloc_aligned` call will fail if
given a layout whose size is not a power of two.

Bug: 274441673
Test: atest MicrodroidTests
Change-Id: I37d7a2805aa6788d22c2cd5bfc8c701fad1f007d