[rialto] Add communication protocol library for host and rialto
This cl adds a communication protocol library that includes
Request and Response definitions for communication between the
host and rialto. The Request and Response messages are
(de)serialized in CBOR format and exchanged via vsock, enabling
the exchange of long messages and the handling of partial packets.

In addition, an end-to-end test has been included to cover this
functionality.

Bug: 291732060
Bug: 292080257
Test: atest rialto_test
Change-Id: I5f6412949e34b2431d060703e6dea1b96c92fde5