PVM to use Secretkeeper for managing RP secrets
The rollback protected secrets are to be stored in Secretkeeper which
provides tamper-evident storage for pVMs.

Regular binder proxy cannot be transferred over RPC binder, so we
SecretkeeperProxy service that forwards the rpc binder request from with
pVM to Secretkeeper HAL which is a regular binderized HAL. This proxy
service is hosted by virtualizationmanager.

This patch hardcodes is_rp_secrets_supported to true to enable testing
VmSecret - v2.

Test: atest MicrodroidTests#encryptedStorageIsPersistent
Bug: 291213394
Change-Id: I3adb78b5eb9d5e7d53b2f990616668e0ceb63471