PVM to use Secretkeeper protected secrets
These secrets are to be stored in Secretkeeper which provides
tamper-evident storage for pVMs.

Regular binder proxy cannot be transferred over RPC binder, so we build
SecretkeeperProxy service that forwards the rpc binder request from
within pVM to Secretkeeper HAL which is a regular binderized HAL. This
proxy service is hosted by virtualizationmanager.

This patch hardcodes is_sk_supported to true to enable testing VmSecret
- v2.

Test: atest MicrodroidTests#encryptedStorageIsPersistent
Bug: 291213394
Change-Id: I3adb78b5eb9d5e7d53b2f990616668e0ceb63471