pvmfw: config: Treat v1.x as latest version known
Currently, pvmfw rejects config data of versions it doesn't support,
which leads to the pVM boot being aborted. Instead, as the config header
format was designed to be forward compatible[*], if a version v1.x
higher than the latest known version is encountered, treat it as the
highest supported version i.e. read the entries that were supported in
that latest version and ignore any new (unknown) entries.

For the same reason, stop rejecting unknown config data flags being set.

[*]: its first fields (magic, version, size, flags) are fixed and its
     array of entries (of version-dependent size) contains offsets from
     the base (not end) of the header.

Bug: 296207806
Test: atest DebugPolicyHostTests
Change-Id: I7969d0189be48a6d54c85965d0092495b43e1982