Dice Policy Aware authentication: Client side lib
Dice policies are to be used by pVMs to seal the secrets in
Secretkeeper. The policies are such that only pVM with certain dice
chains can access the secrets. The constraints will be set by pVM (see
`keys_to_copy`).

This module introduces the client side code (that will run by
Microdroid) for constructing appropriate polcies.

Test: atest libdice_policy.test
Bug: 291233378
Change-Id: I32b78cefd77a9fd1f62800fd15569aea912f60bd