Remove statsd APEX, clean linkerconfig
We don't actually use the statsd APEX in Microdroid, so remove it.

Also simplify the code that adds the required APEXes, get rid of some
unnecessary Vecs.

I noticed a reference to related libraries in linker.config.json - but
they seem to have no effect. Looking at the history we added this file
as part of b/181093750 when we wanted the ART APEX in the VM. But that
is only needed by CompOS, which works just fine without this. None of
the libraries mentioned here currently seem to exit in Microdroid's
system/lib64, so I just removed them all. (Removing the file entirely
breaks the build.)

Bug: 298195944
Test: Builds
Test: composd_cmd test-compile
Change-Id: Idf82bae9846d5d02371845edf0c6ef4c318b9625