pvmfw: Print RebootReason on secondary UART
When rebooting due to a business logic error, print the reason as a
formatted string to a dedicated UART, in the same way Microdroid does it
on its /dev/tty1, to provide the host with extra information about the
reason for the pVM rebooting.

Since aosp/2646391, all the pvmfw code runs with the UART properly
initialized, which greatly simplifies printing of the reboot reason as
start() can now assume that the secondary UART is also ready to be used.

This CL was verified through the logs of virtualizationmanager e.g. in
MicrodroidHostTests#protectedVmWithImageSignedWithDifferentKeyRunsPvmfw:

    virtmgr : virtualizationmanager::crosvm: VM returned failure reason 'PVM_FIRMWARE_PAYLOAD_VERIFICATION_FAILED'

Note that the host still has to be taught about the new PVM_FIRMWARE_*.

Bug: 220071963
Test: atest MicrodroidHostTestCases
Change-Id: I5d382be9dae02da7e463193400dfa0474ca5e378