pvmfw: Rollback index of kernel & security_version
Capture rollback_index of guests' kernel. Rollback indexes are available
in AvbSlotVerifyData returned from avb_slot_verify(). This is a slice of
uint64 where the position of the rollback_index is determined by
rollback_index_location (which defaults to 0).

Then this used as the 'security_version' in the dice_config_values that
can be used to provide rollback protection to secrets.

Note on TrunkStableFlagging - This is guarded by flag llpvm_changes,
based on which security_version is added to  dice.

Test: #payload_with_rollback_index
Test: flash pvmfw => get dice chain from Compos => verify-dice-chain
contaisn security version = 1 for guest OS
Bug: 296830692
Change-Id: Ia10519eb610b487d20a2448f232c0e82cb95e730

Change-Id: I0d6d993d8b2d1b98dcc39fb90895a59c7a699d7d