Add guest OS capability: SecretkeeperProtection
Not all guest OS are capable of interacting with Secretkeeper. Add a
capability "secretkeeper_protection" (which can be extracted from vbmeta
property "com.android.virt.cap"). Add this property to Microdroid
kernel.

pvmfw will have check if the guest OS has this capability & ensures the
rollback_index > 0 if the guest OS has it. Note that this will be
factored in while pvmfw check if updated guest OS should be accepted.

Bug: 29121337
Test: avbtool.py --info microdroid_kernel & check if property is
present.
Test: #payload_with_multiple_capabilities
Change-Id: I2ae3de06b4bf7a277071c69193ca9e714c1305bd

Change-Id: I99c159d3d65005ec02729b47620ac05ab8d1ec5e