[client-vm] Add certificate request skeleton code for client VM
This cl introduces a new request/response type for the client VM,
enabling it to request attestation from the service VM. The
client VM sends the request to the service VM and receives the
corresponding response type back from the host.

The specific details and handling of the request will be
completed at a later stage.

Bug: 278717513
Test: atest MicrodroidHostTests
Change-Id: I7c073660a3cf7e2ecb07a9239b1dad7c9f3afe00