secretkeeper_comm library: for Sk & client impl
The purpose of the library is to serve as helper library for
Secretkeeper HAL implementation as well as the client stub.

In this patch we implement (subset of) data structures specified by
SecretManagement subset of the HAL. In particular, we implement
GetVersionRequest, GetVersionResponse and SecretkeeperError data types.

Also included are the units tests for serialization/deserialization.

Test: atest libsecretkeeper_comm.test
Bug: 291228655
Change-Id: I33229867817eb45611c4db9b2e3005bc75670be0