Revert "Skip changingDebuggability related tests on Non-pVM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41208e95-668f-4508-b1bf-45c519183288). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3184525c4450f6223da545edac939c9ed57600ec