Revert "[bssl] Remove redundant libcrypto_baremetal in libbssl_a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c32f8563-230f-421f-a4d9-dc1cd7c92404). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02c37a8ffad248292b4924e077489138af43dc17