Revert "[bssl] Remove redundant libcrypto_baremetal in libbssl_a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21a9ead7-cf69-4f2d-9c2e-c852f9c1b4df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I34a4781c8b0e74a6dc056c01924ef9ac36457dbb