Add APK details to the DICE chain
Include the information from the manifest in ApkData (which also means
it is persisted in the instance image - which is not necessary but
harmless).

Add a CDDL file descibing what the VM config descriptor looks like.

Add information about the APKs to the Microdroid payload configuration
descriptor.

Also make some formatting fixes to open_dice error messages (since I
managed to trigger some).

Bug: 299591171
Test: composd_cmd test-compile; manually inspect BCC
Test: atest MicrodroidTests
Test: atest microdroid_manager_test
Change-Id: Iad7e8407cd9ad1d6715806aa4479641b9b9173cf