Revert "Upgrade zerocopy to 0.7.5"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9f3ffee-9171-403d-9de0-efee020ae68c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4bfcdd2fcbd9bd6c186e7a893dabf8115473319