Revert "Prepare for keeping concrete overrides of abstract methods"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d6c76e4-1db4-43d1-b563-89b31019216b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie38089b484ded52f2ddec2f4b5ef7d79123b0d26