Refactor Dice policy test
A substantial portion of unit testing dice policy structure is about
creating test dice_chains, constraint_spec & expected dice policy.

Move these to a separate struct with init() method, so that unit tests
could directly use this TestExample. As we will see, it will be useful
as we need for dice chains for more comprehensive testing.

Test: atest libdice_policy.test
Bug: 291238565
Change-Id: Ie9b96e231ded334cc2cb370aa8a320a8a5e9c757