pvmfw: Replace const_cstr! with now-const cstr!
Now that vmbase provides a const-compatible macro (aosp/2819377), there
is no need for const_cstr! or its const values (as the macro can now be
used in const contexts) so improve the readability of the code (with no
runtime cost) by inlining the cstr! calls as if they were literals.

Test: m pvmfw_bin
Test: m libpvmfw.device_assignment.test
Change-Id: Ie2ad127ae4181ef4c073b59db4525852206b73c7