Enable verifying vendor partition
DO NOT MERGE (yet)

This change would make the vendor partition be verified by public key in
DT node. The key is passed from the host for the non-protected VM, and
from the pvmfw for the protected VM. This change should be submitted
after supporting passing key for the pvmfw.

Bug: 285855436
Test: N/A
Change-Id: Ic9de35509a42ecdc7b00cc2ec3dd2dbcc1d71125