[service-vm] Deserialize CBOR-encoded CSR byte array into struct
This CL deserializes the CBOR-encoded CSR provided as a param for
VM attestation into a `Csr` struct.

It also extracts the CSR/attestation key generation function from
microdroid into a standalone library so that we can reuse it in
e2e tests.

The new test target has been added to the busytown config at
cl/581170242.

Test: atest rialto_test
Bug: 309440321
Change-Id: Idbfadd0418822bdf46d6764b9db484f4865aa267