Flag guard removal of the vendor partition from Microdroid super
The
MicrodroidHostTests#testBootFailsWhenVbMetaDigestDoesNotMatchBootconfig
test built as part of Android UDC CTS binary will fail when running
against the -next targets, since that test excepts existence of the
vendor partition.

This patch fixes the test by conditionally removing the vendor partition
(on -next builds it will still be present). Note that the /vendor still
won't be mounted, since that change is not flag-guarded. I think this
should be ok, because in Android U the /vendor partition was effectively
empty.

Bug: 307904471
Test: atest MicrodroidHostTestCases
Change-Id: I40fa72f1343c881bdda546f7539303f2ff971e55