virtmgr: crosvm: Support VFIO iommu=pkvm-iommu
To permit the transition away from the hack (ag/25008183) which was
hijacking VFIO_TYPE1, detect if the currently running build of crosvm
supports the proper pKVM pvIOMMU type (crrev.com/c/5009546) and only
fall back to the wrong/hacked VFIO_TYPE1 otherwise.

Leave a TODO to replace this logic with the proper value hardcoded once
we've all moved to builds that support it.

Bug: 306110430
Test: ehdemo/run_in_microdroid.sh # With/without support for pkvm-iommu
Change-Id: I5db4c8533d5dc1ca3237f2a9182ea6b83151ea57