Revert "Add GKI kernel to virt APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6577be87-0e96-4965-8337-88ad2a78fb09). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I097e60e0c02ecd68047bba1949b5302300d9e362