Revert "[x509] Parse certificate chain from RKPD into individual..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f460d71-0638-4132-b52b-73ab967ae9ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I545ea0a1faa266d29a455713bd81e7125bd02327