Compress Microdroid GKI kernel with LZ4
Reduce the size of the com.android.virt APEX by compressing the GKI
kernel with LZ4. The downside is a slightly longer boot time.

Note that we compress the kernel AFTER it has been signed. That's
because the binary blob will be decompressed into guest memory first,
and only then verified by pvmfw.

Bug: 315141974
Test: atest MicrodroidBenchmarks
Change-Id: I6cb1ae33f9572a4592a04d3163a933d41d2566fe