Add Secretkeeper key support
Create a new version of pvmfw config that has a config entry for the
Secretkeeper public key. Add it to the DT if present and the llpvm
flag is enabled.

Also a whole lot of further refactoring to make adding new config
entries a little easier, and to generally simplify.

I also removed the mut from some of the config entries. And lint
nagged me to introduce an FdtConfig struct rather than add yet another
parameter, but I think the result is a little neater.

TODO: Update the README, define the key format, figure out the max key
size.

Bug: 291232226
Test: flash modified ABL that populates the key (gpar/777513);
  start a VM with vm_shell;
  cat /sys/firmware/devicetree/base/chosen/avf,secretkeeper-key;
  See the expected output.
Change-Id: Ia699c27bd669644c7a054c88f6ad4f2923a20356