Revert "[attestation] Validate client VM's Microdroid kernel in ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68027675-360c-46a2-bf47-f60004624d64). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I044d0a28315dcd758ff996971d253d0f0481775e