Revert "pVM to use Secretkeeper protected secrets"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/281662ae-9001-4ab2-b1e6-497ca83ed760). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6591629c734f5797101afeeaf1a679de3f18c030