Revert "[pvmfw] Use AEAD wrapper from libbssl_avf"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c356dfa1-adda-451d-8e01-0b1c53a3b1c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1626f16d79ea28781250b89cb53a3936c916d012