[test] Add e2e instrumentation test for VM attestation
This cl adds an e2e test to check the VM attestation with the
following steps:

- Provisioning a pair of mock keys in virtualizationservice.
- Requesting VM attestation from a running VM and tracing this
step.
- Validating the attestation result.

To facilitate the testing process, two new APIs have been added.
The first API allows for the provisioning of mock keys, while the
second API facilitates the request for attestation in testing mode.

The original test_apk is renamed to demo_apk, as it will serve
as a demo app that interacts the same way as a client app.

Bug: 318333789
Test: atest AttestationClientTestApp
Change-Id: Icf9d39c704af316f7625cf9d057d255243eca445