Add a VirtualizationSystemService that runs inside system_server
This service will later be used to subscribe to trigger clean up
routines when an app is deleted.

The code is located in the
/apex/com.android.virt/javalib/service-virtualization.jar. This jar
contributes to the system_server classpath.

The service-virtualization.jar will be installed for both AVF enabled &
AVF disabled variants of the virt APEX. However, on the devices that do
not support AVF (i.e. using AVF disabled virt APEX), the
VirtualizationSystemService won't be started, since the AVF disabled
virt APEX won't have the `<apex-system-service>` entry in it's
AndroidManifest.xml

The change is guarded by the RELEASE_AVF_ENABLE_LLPVM_CHANGES build
flag.

Bug: 323339759
Test: boot device & check logs
Change-Id: I948be5eb38a9911d510f63e4976c117ccdb291f8