[VirtMgr] Allow addition of avfnonsecure properties
Allow addition of properties at /avfnonsecure in the DT overlay created
by virtmgr. While I'm here, refactor referece_dt into dt_overlay to
allow this & remove confusion of reference DT which from the pt. of view
of pVM is expected from ABL.

Include (a hardcoded)`Id` of a VM in the the DT overlay!

Test: vm_shell start-microdroid --auto-connect
Test: cat /proc/device-tree/avfnonsecure/Id
Test: atest virtualizationmanager_device_test
Bug: 291213394

Change-Id: I4ffa8a051532714bba091e7112e757e8b3b651fc