Remove Microdroid tests related to MANAGE_VIRTUAL_MACHINE
With associating a new gid for MANAGE_VIRTUAL_MACHINE permissions,
this causes ActivityManager to kill processes that attempt
to modify its gid at runtime.

Remove runtime grant/revoke of MANAGE_VIRTUAL_MACHINE permissions
from existing tests. For the tests that are testing for a lack of
MANAGE_VIRTUAL_MACHINE, those can be split off to a different apk.

Bug: 322197421
Test: atest MicrodroidTests

Change-Id: I1ec62228462efb198ada9141ad91c987d0f8710d
Signed-off-by: David Dai <davidai@google.com>