Allocate each VM an instance_id
Introduce a 64 bytes' long instance_id. When the VM is created, this is
requested from virtualization service, which randomly allocates it.
While it does that, it also logs the user & the package name (It gets
package name from Package Manager, this will be later persisted by VS in
a database).

The app persists this allocated instance_id in a file `instance_id` in
its directory for the vm instance (along with instance.img &
storage.img). When the VirtualMachine is run, this is an input into the
VM via DT.

This patch modifies Compos & /vm/ binary to work with the instance_id.

Note on flagging: Inclusion of instance_id in Guest-DT is conditional to LLPVM feature flag
Bug: 291213394
Test: atest MicrodroidHostTest
Test: atest MicrodroidTests
Test: atest ComposHostTestCases
Test: Looks for instance_id logged by VS
Change-Id: I2f58ea0c06b2d21a90aae5f6fc45c8e09de6b3a7

Change-Id: Ie8e25b9510e27362d4580c55c1bd557143ff7d0e