Revert "More javadoc for capabilities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c22e4a6c-54af-46ee-8352-ac107581bdfa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e49fb4f9d000ace50de01bdead8f9fe4c54bec7