Revert "Allow clippy::unnecessary_fallible_conversions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7368812a-b82d-45cb-bbc9-a59b78369d11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a7355b1d9b21b74b0fa63d5b8558fc727f6c65a