Microdroid: Skip instance img if sk is supported
Microdroid no more needs Instance image partition if Secretkeeper is
enabled.

The use of instance.img is to store package data at first boot of the
instance & MM ensures that it did not change on further boot.  With
Secretkeeper based rollback protection, the auth_hash & version of each
of these packages are part of DICE Policy & Sk ensure the secrets are
not released if the version downgrades or auth_hash changes.  Therefore,
there is no longer any need for this data to be in instance.img

Note: Since Secretkeeper is an optional HAL in Android V, we still need
to support the instance.img for cases when Sk is not present.

Bug: 291306122
Test: TBD: hexdump instance to see Microdroid partition header is
present

Change-Id: Iad97f2c81f759a1c13eac6d1302dee43be534e11