Remove odsign_e2e_tests_full from avf-postsubmit group
The test should match with the ART implementation in DUT.
The implementation can be built from ToT in case of trunk,
but could be a prebuilt from mainline-train-builds in case
of next.

Until we find a way to use different tests depending on
the original of the ART module in DUT, drop the test
from the avf-postsubmit group.

Bug: 325409306
Bug: 290882298
Bug: 326970965
Change-Id: I1734c8df44952f78e5b2f158c711bea95dd50d78
Test: watch treehugger