Revert "pVMs to use Sk for rollback protected secrets"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edac43bd-4212-4e12-92ee-f60c8f40d27f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe86cd323f223a7c168a997faf77589a7a796d27