pvmfw: Get instance salt from instance-id
Make salt in instance.img obsolete. It should intstead be derived from
instance_id in DT.

As far as `new_instance` is concerned, if instance.img is present
(becuae Sk is not supported on device), instance img can still be used
to determine it. But if Sk is supported on devices (defer_rpb is set)
new_instance can no more determined, let it be false. Microdroid can use
check if the get_secret() returns EntryNotFound Error & that would
indicate that this is indeed the first run of the instance.

Test: TBD
Bug: 291213394
Change-Id: Iebfcdd5d89513b8f24c937b2f2cd9a298d01b74d