Revert^2 "Make salt in Microdroid's instance img obsolete"
Salt has been used to provide differentiation of secrets of 2
non-protected VMs (as hidden input in DICE). Have hidden input be
derived from instance_id of the (non protected) VM. In all other cases,
it will be all 0s.

Test: Microdroid tests use this code path.
Bug: 291306122
Re-revert: The reason for failures was a different issue(b/291306122),
this patch is harmless.

Change-Id: I86727ce5361b2cf4e9be59b2f8c0c5e76af9d2e8