Revert "Allow clippy::unnecessary_fallible_conversions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fc2a66e8-2f18-4bfd-b215-588641cfed50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb24d91b65c01a207cd5e08f6858adea3f88e43a