[instanceId] allocate instance ID for the Service VM
This cl allocates an instance ID to the service VM since the
instance image has been deprecated. The instance ID is
used in pvmfw to deduct the sealing CDI needed by the
Service VM request processing.

An instance ID has also been added to the non-protected
rialto in rialto_test in this cl.

Bug: 327593612
Test: atest VmAttestationTests rialto_test
Change-Id: I939957f0be5ae33a6205fd55819e715ae2792b87