[attestation]  Verify AVF RKP Hal presence in VM Attestation
This cl adds a verification step to check the presence of AVF RKP
Hal during the VM attestation. This step ensures that in cases
where a device lacks support for VM attestation, an
AVmAttestationStatus: ATTESTATION_ERROR_UNSUPPORTED will be
returned to the VM.

This return status will also be used in subsequent tests to skip
the test if pVM remote attestation is not supported on the device.

Bug: 329652894
Test: atest AvfRkpdVmAttestationTestApp
Change-Id: If2bffde34dcd2a8bb78b9aa06f7aefc073f427d6