compos: Enable huge_pages
The huge_pages VM option (a.k.a. transparent huge-pages in kernel land)
gives a hint to the kernel the guest memory can be backed with
huge-pages. (i.e. 2MB on 4K arm64 systems).

On a Pixel6, this lead to a 2 seconds improvement, out of a total of 30
seconds for test-compile.

Notice although this will have no effect, until
/sys/kernel/mm/transparent_hugepages/shmem_enabled is set to "advise".

Bug: 278011447
Change-Id: Ic3da81c0ae7777a4954d4cbd9ab7e8a0ac7652c9