Revert "Set --lto-O0 to reduce binary size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ca33e3f-2ef9-4f41-bdfa-4dd0d70e56f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b8c410fbf481ed34951ae11edc1679f59cf9c3a