Revert "Set --lto-O0 to reduce binary size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39d03588-4687-4e6d-aed6-32d4b46b8f0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfba7e10b30cf8577d371be169bb308f0bca6075