Revert "Set --lto-O0 to reduce binary size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e05a9ddf-3bba-4e68-94b9-63ee161bf217). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e81dd9c43b7f0ca0c3aeb97db86b37722174506