Revert "Set --lto-O0 to reduce binary size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69b9aa60-7b29-43f6-93cc-d789661ba111). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99bba99f27a214412e152d69ce604b398cdcc0bc