Revert "Match with the recent TEST_ROOT changes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4103c4a9-0f1f-449c-ba2f-3604c52784a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4f888b04c24ff788a0e95c7e5d5d95c96aeb5c1