Revert "Set --lto-O0 to reduce binary size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82e46832-28dd-49ee-af15-451053fa920e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c482b6bd3971b3e4b50b816e03a32d806120b79