pvmfw: Unify Config::new() calls into a single one
We currently parse the config data twice due to a bug as a safe way to
bypass a bug in the Rust borrow checker which was causing the 2 mutable
borrows of data to be seen as overlapping (when they don't). However,
this causes unnecessary work and, perhaps more importantly, creates
duplicate logs.

Instead, implement the match in the most straightforward way (as
originally desired in aosp/2345203) with a Ok/Err match and avoid the BC
getting confused by carrying data into the "Err" branch by temporarily
removing any ownership information from it (unsafe but clear).

Test: m pvmfw && TH
Change-Id: I170e6b565266df7f0ed43c41ce1f397f089c03bd