DO NOT SUBMIT Write vsock benchmark in bash
Unfortunately, it is quite hard to capture VM traces from the current
java benchmark, hence this patch that adds the shell version of the
benchmark together with some hacks to make it possible to run it against
multiple kernels. In our case the kernels are:
* microdroid (based on android15-6.6 branch)
* microdroid_14_61 (based on android14-6.1 branch)

I do not intend to merge this patch.

Bug: 333701168
Test: run vsock_debug.sh
Change-Id: Iab313ef6b464975cae516422b28bc9a5df0a7a55