virtualizationmanager: use raw file instead of qcow
crosvm's qcow backend isn't as well maintained or optimized as its raw
file backend. The OS should lazily allocate blocks in the filesystem as
they are used since we are only setting the file length (i.e. ftruncate)
and not requesting it to be allocated eagerly (i.e. fallocate), so the
disk usage should be similar to using qcow.

Test: TH
Change-Id: I4e456314736d5a99b850904afc3f1047d6c2e885