Guest OS serial console
Create a pty on host and forward the guest VM serial console.
A background thread would be submitted to do the data polling. It is the
APP's responsibility to shutdown the thread and close the pty file
descriptor after the VM is shutdown.

After the VM is started, the serial console name would be printed to the
logcat (/dev/pts/<num>). The number of the pts node is not fixed as it
is dynamically allocated by the pty driver.

You can then connect to the serial console through adb:
```
adb shell cat /dev/pts/0
adb shell -t microcom /dev/pts/0
adb shell -t 'stty raw -echo && netcat -f /dev/pts/0'
```
Keep in mind that special characters like ctrl-c and ctrl-z are _not_
intepreted, so you would not be able to interrupt the host netcat
command by typing ctrl-c. Pressing ctrl-c would send the ctrl-c keycode
to the other end of the serial console, and cause an interrupt in the
guest VM console.

If you want to be able to interrupt and stop the host netcat command,
pass the `isig` terminal option:
```
adb shell -t 'stty raw isig -echo && netcat -f /dev/pts/0'
```

All output data from the guest serial console is also written to
/data/user/0/com.android.virtualization.vmlauncher/cache/console.log
so you can retrieve them later in case you lost your serial console
output, or if your terminal window becomes garbled with junk.
The file is not world readable though, so is only retrievable on a
debuggale host build.

BYPASS_INCLUSIVE_LANGUAGE_REASON=It is well known that "man" stands for
the "[man]ual" command in the unix world.

Bug: 335362012
Test: Start a vm and then `adb shell -t microcom /dev/pts/0`
Change-Id: I9179510b237b8a19b8da7c527f3f42846a127dd3