Make ecdsa_verify_nist testable
Add ecdsa_sign_nist (and rename the existing ecdsa_sign to
ecdsa_sign_der). Implement DER to NIST signature conversion to support
this.

Add tests to ensure that signing & verification are compatible.

Note that while I'm adding code to the production bssl library, none
of it is currently used other than in the new tests, so this shouldn't
break anything.

Bug: 338745127
Test: atest libbssl_avf_nostd.test
Change-Id: Ie522f556f90c4515b4f738a23a42c02ff48b073e