Revert "virtualizationmanager: use raw file instead of qcow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a63f1199-b11a-40a2-bf52-04d52e1c4ca3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55da0c93c8938209a8b55d701575762e371c0547
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com