Use soong variable avf_remote_attestation_enabled to enable RA
The new soong variable allows OEMs to disale the RA feature
easily when they do not plan to support it.

The remote attestation feature will be disabled if
PRODUCT_AVF_RA_ENABLED is set to false in the make file,
otherwise it will be enabled when the feature flag
RELEASE_AVF_ENABLE_REMOTE_ATTESTATION is on.

Bug: 331269709
Bug: 341292601
Test: run VM remote attestation on Pixel 6 and check it is
disabled

Change-Id: Idcc1c4c300f64661293394e7102cf68c3bdc38f2