Use soong variable avf_remote_attestation_enabled to enable RA
The new soong variable allows OEMs to disale the RA feature
easily when they do not plan to support it.

The remote attestation feature is opt-in by default, as it
is a strongly recommended feature from Android V. It will
be disabled if PRODUCT_AVF_REMOTE_ATTESTATION_DISABLED is
set to true in the Makefile, otherwise it will be enabled
when the feature flag
RELEASE_AVF_ENABLE_REMOTE_ATTESTATION is on.

Bug: 331269709
Bug: 341292601
Test: run VM remote attestation on Pixel 6 and check it is
disabled
Test: run VM remote attestation on Pixel 7 and check it is
enabled
Change-Id: Idcc1c4c300f64661293394e7102cf68c3bdc38f2

Change-Id: I899c670b62da7cc76b61c27f978b3e2164e100b1