Revert "Re-enable AVF capability test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcd8ca93-8793-4a06-8310-27cb11f29446). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341881508
Change-Id: Idc860703f4460139ef2dd3afe3d00bfd7412be70
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com