Revert "Allow `null` kernels in VirtualMachineConfig"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e3c36d09-aaec-4808-8faa-66496aad77d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I269529daedc76d34417f3bab086aa5ac50bb42ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com