Revert "Use soong variable avf_remote_attestation_enabled to ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/866bd2cc-77cf-415f-b6a6-1fe33e0e0e78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53b949a2ccc19879a4379f6f224fdcafed9e8164
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com