Revert "Use soong variable avf_remote_attestation_enabled to ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12824617-026a-46ec-885c-d33dbee30ce9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifed2e4fc20bcde172e9daf80ad039ace1b08b870
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com