Revert "Use soong variable avf_remote_attestation_enabled to ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c85167e-011f-4bc2-9112-b8f1788c8e3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c47abe09e96959da9d25a6e591d3c4f28781301
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com