Revert "Use soong variable avf_remote_attestation_enabled to ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcf7c203-eb2c-487f-bfe7-784c9ef96c6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fadc937904468cb64704a49c9069c8484103367
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com