Moving rust openssl use in virtualizationservice to bssl-crypto
Migrating the rust openssl usage in this module over to
bssl-crypto.

This is currently failing the test splitting_x509_certificate_chain_succeeds
(the test in src/aidl.rs, which tests the changes here).
This CL is a WIP to show the changes and hopefully get some help figuring
out what the problem is.

Test: atest virtualizationservice_test
Change-Id: Ifd5e4f3a2b7efd6f2bcd62a303fcc864e913809c