Revert "virtmgr: Add boost_uclamp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c41f4d48-1789-47c9-817e-24d7eaecab84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347203579
Change-Id: I044e68fb077634a298258d2d0d89aac8d8beb5da
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com