Revert "Skeleton for IVmTethering service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3c6823e-6fbe-4a97-b810-1764e1b4f249). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347407635
Change-Id: Ie9b263d7b2e1e774f26ef64c1ef5df4bd753dd63
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com