Revert "Small fix of type conversion of TUNGETIFF in vmnic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ca77b5b-9b53-457f-87db-d38ae3c22b8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347604033
Change-Id: I5dd97949a22c67929ed4a6ca6312c954937efec2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com