Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a6575b2-e582-4694-b3a3-cbc66a180df2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bbf43318f772e6f6f7c5593a5376361e45a7c0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com