Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fc388080-9fd3-4ca7-8378-1803d156aad6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iacc6f55661bc43940b4e6f93ad7270ba55cde908
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com