Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/06c3073f-04d3-4112-b62e-8af587847afc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8bea428b8420bed201d1b143bb3b21f8cb7b3f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com