Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a135d184-9687-4401-a85a-a74e785d007f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2691a219b5540eefb5ec7b7121638ead483035d0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com