Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/085d7910-b96f-4166-93aa-b759f9f85c37). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26b0967036b4687e7a4edc0128e6ea74d8e5f1fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com