Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e02b85bc-7510-4629-8d74-ee347e8c8db1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347862795
Change-Id: I34752f50496958cc4ed2c93fe99fee3d7ec166a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com