vmbase: Improve safety of console::init()
As the console should only be accessed once it has been MMIO guarded and
mapped in the MMU, prevent the emergency_write_*() from writing to the
UART before console::init() has been called.

Mark console::init() as unsafe, documenting the constraints on when it
should be called and, in the entry code, move it _after_ we've called
mmio_guard.map(UART_PAGE).

Move BASE_ADDRESS to the more appropriate layout module and allow the
caller to pass the UART base to console::init().

Test: m pvmfw librialto libvmbase_example
Change-Id: I9543765fd0429b38b45bc2e7b7c83bedf79194f3