vmbase: Support secondary UARTs 0x2f8,0x3e8,0x2e8
Parameterize the code to allow printing to the "n-th" console and
setting up an arbitrary (up to MAX_CONSOLES) number of UARTs to be used
by client code.

Extend the layout to describe the 4 UARTs crosvm unconditionally
generates and configure console::init() to allow printing to any one of
them.

Test: m pvmfw librialto libvmbase_example
Change-Id: I899f70dfb5eeb41e4ba399aee59d731db1264ff9