vmbase: Replace emergency_write_str with eprintln!
Similar to vmbase_example and pvmfw, use the macro in Rialto, instead of
the function directly.

Update the README to only recommend eprintln!() as the comment about
emergency_write_str() not accessing the stack is not guaranteed (and
relies on an implementation-defined behavior of the compiler), which is
less likely to be relevant since aosp/2550864 as synchronous exception
handlers now have a dedicated stack.

Test: m pvmfw librialto libvmbase_example
Change-Id: Ia2c90b0c2896016fa93a0f2cd326ae72844a9c84