Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84367bda-ba49-40ba-9c03-c32602d56072). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53afb95b8e209a64b31c7a17280ddb3d78e80aa2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com