Revert "Ferrochrome: Don't set kernel"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c164a68d-8ff8-4fee-9923-9cdf1bf58742). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e0280d00b0d3acafe6ab0d417881f1ca4ae4ad0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com