Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02c3a387-6c98-4cfb-9cff-6deb8e6f1047). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347861095
Change-Id: I60d6f1d7a91b443bbb50f72e5ff717059759deba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com