Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6c4dd4a-92d4-4d61-b818-7cddf2435d6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347840973
Change-Id: I6674076b5894841addad839dc0b8500fdfcffcd8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com