vmbase: bionic: Make FILE* non-dereferencable
In C, it is legal (albeit impl-defined) to dereference a FILE*. However,
treating our CFilePtr variants as valid pointers would be unsafe (from a
Rust perspective) as they are not actual (Rust) pointers. Currently, we
have no need for an actual FILE struct so there is no proper way to
construct valid pointers.

Instead, ensure that C never tries to dereference the FILE* values we
pass it for stdout and stderr. On AArch64, the most robust way to do so
is by using the "middle" VA range, which the MMU is unable to translate.
As a result, C dereferencing these pointers would trigger an exception
leading to a system crash, keeping the Rust environment safe.

However, as those addresses use more than 32 bits, Rust warn about them
not being portable as usize. Therefore, enfore that the code be only
compiled for AArch64, rendering moot the portability concern, which can
then be silenced. This approach will force other targets to re-evaluate
the assumption and/or define their own addresses.

Test: m pvmfw librialto libvmbase_example
Change-Id: Ic4e373fa86db03728f92cd522fdb089462c11f83