Revert "Add Virtio-snd device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675dbfd5-bf20-4732-a864-369731712e82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia511f1dcd8656ef3deb5f3abe3d8104127f1a20c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com