Revert "Microdroid: Add support for android15-6.6 GKI as guest k..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b83e02c-bd65-411e-b06e-548e52668466). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb3c7e18f5283396e491a33136803eb6cbf6e3aa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com