Update patch set 3
Patch Set 3:

We were unable to auto-submit your change due to the following error: Failed to submit 2 changes due to the following problems:

Change 3152720: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Change 3152721: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.



 [view details](https://android-build.googleplex.com/presubmit-status?change_id=platform/packages/modules/Virtualization~3152721&revision_id=3&host=android&id=6574210174779392)

Patch-set: 3
Tag: autogenerated:TreeHugger:Autosubmit-Fail
CC: Gerrit User 1974282 <1974282@85c56323-6fa9-3386-8a01-6480fb634889>