GUID of a partion can be specified if needed
Previously, when a composite disk is create from one or more partition
images, the GUID for each of the partition was randomly chosen.

This change makes it possible to specify the partition GUID if needed.
This is required for Ferrochrome where partitions are accessed via their
GUID, not by their label in the partition table.

Bug: 351973725
Test: Ferrochrome boots with the vm_config.json in the follow-up change.

Change-Id: I63647d77d74a7be99cc4811f5c836f74f13928a9