Use a message handler to offload input events
Instead use an handler that listens for messages and only
then send the events down the wire to the VM. It should
be equivalent to the old implementation, minus being
more typical in Android as a mean of communication of
available data.

Furthermore the methods used to send events now can only fail
if the target is null or the thread is dead, this shouldn't
be possible as we always have a target and the thread is shut
down after the VM, so remove the boolean return type and make
the caller manually return true as the event will be eventually
digested.

Test: boot ferrochrome, use keyboard/mouse
Change-Id: Id360d15ceee83ebb891eb521522d003b5412b8f8