Use a message handler to offload input events
The current implementation actively waits for
new events to be added to a queue, this is especially
problematic in case there are no events and the loop
is busy try/catching exceptions causing uneeded CPU usage.

Instead use an handler that listens for messages and only
then send the events down the wire to the VM.

Futhermore the methods used to send events now can't fail
as the events are simply added to a message queue, so make
them void and make the caller maually return true as the
event will be eventually digested.

Test: boot ferrochrome, use keyboard/mouse
Change-Id: Id360d15ceee83ebb891eb521522d003b5412b8f8