Revert "Use try-with-resources mechanism for handling clipboard ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59bfb847-6373-4e28-9e7d-f6b95eb3a256). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id61b3f1522179b3ec5aec0e6b25cf666021f1572
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com