Add FerrochromeApp as a share target
Allow sharing text data to FerrochromeApp, if text looks like an URL
(supports http / https / mailto link for now), open the link in VM.

Add a new OpenUrlActivity in FerrochromeApp to handle this feature.
OpenUrlActivity is launched with the 'singleTask' launch mode and this
ensures it is instantiated in the same taskAffinity task as the
FerrochromeActivity. It then bounces the share Intent to the already
launched VM activity.

Change launch mode of VmLauncherApp to 'singleTop' so that we can send
Intent to the already launched VM activity. If we don't use 'singleTop'
launch mode then we would instead create multiple instance of the VM
activity. When the VM activity receives a intent via onNewIntent(), it
sends the Intent payload to the VM data sharing service.

KI: VM need to be logged in first for link sharing to work.

Bug: 348303697
Test: 1. Launch FerrochromeApp and login
  2. Switch to another app to select a link
  3. Share the link to FerrochromeApp
Change-Id: I39742d4e897caaa37595f50697540196d0393946