Revert "Add com.android.virt.accessor_demo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13b914d2-d226-4a8c-93dc-b6c1b053a600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7a5926b577438534b1bec5ff95618b8b11bd2ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com