Revert "Add com.android.virt.accessor_demo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbac161-b81c-495d-9fab-0215156ae502). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5ded1046da49f8f0fc31492d38c15407bb134b4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com