Revert "Add com.android.virt.accessor_demo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fd2c5ea-42e1-42a0-847d-41036ed8d048). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26568e75ae760540657ce06674a5ce6c28af6214
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com