Revert "Set tablet mode if there is a physical keyboard"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d4fcf21-857e-45cf-8f2d-1cf00c611b7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia04ec912d56914a37bb69e2886bb48bd7e5ae7ae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com