Revert "Add com.android.virt.accessor_demo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d739370f-1264-4fb6-bffd-cafedf3d2fa9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibdefed77ef424f15d3e3ee423137dfbeef82c699
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com