Revert "Use build.prop instead of buildinfo.prop"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1037d00d-9c47-469e-bc28-70116f295d35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 355330138
Change-Id: Id347199f50ae19c6276a56498e8c9334d2aae4c5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com