Pretend that pVMs are supported on Cuttlefish
Currently Cuttlefish does not support starting pVM within it. For our
testing purposes, we would like to be able to run applications (e.g.
CompilationOS) that request pVMs. In this change, we add support for
running unprotected VMs as pVMs, within Cuttlefish (AVD).

We check the `ro.product.device` sysprop and only for `vsoc_x86_64`,
we start an unprotected VM in CrosVM, when requested to start a pVM.

Bug: 353716897
Test: cvd start -extra_kernel_cmdline="androidboot.hypervisor.protected_vm.supported=1" && \
      adb shell /apex.com.android.compos/bin/composd_cmd test-compile

Change-Id: I8e0f78eda18f5af7d51e0f898abf80a0f770d831