Revert "Use build.prop instead of buildinfo.prop"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94a813b9-c03e-41a7-bea2-8f1800258032). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I311e524c287af0b81fa0f83d5c7460329bcb9c96
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com