Revert "Use select() for build/apex/Android.bp modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94a813b9-c03e-41a7-bea2-8f1800258032). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2fc5197b9a2395ea32aca6ec9717b068d8c60669
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com