fix rialto on gunyah
Gunyah currently requires the payload to be loaded at the start of RAM,
but crosvm puts the FDT at the start of RAM when using `--bootloader`.

Work around this by adding a VirtualMachineRawConfig option to opt-in to
a different memory layout, with the payload at the start of RAM and the
FDT following it. I'd prefer this to be the default, but more is
required to fix all the users.

Bug: 355696053
Bug: 353758405
Test: atest rialto_test
Change-Id: Icddadf39bee455d11f64a63da0a250a610d5321a