Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91815f2d-427e-4a82-9317-b685988aeccf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 356856433
Change-Id: I3ee1085b6df28071ffb68a10c1feb951d599e894
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com