Revert "[pvmfw/rialto] Disable full LTO in vmbase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/315d3cfb-ce40-4e43-bb04-fda3aeb2c77c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 356856433
Change-Id: I2739de089aaab836536a5255f1e4278e6d871d48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com