Add an example for VM to host servicemanagement to microdroid tests
Microdroid VM payloads will now have access to the VINTF stable
android.frameworks.stats service.

This adds an example to vm_accessor_test where the
vm_payload attempts to get a binder for an host service.

Virtmgr on the host acts as a proxy since it is able to get the host
service through kernel binder and create VSOCK servers.

Test: atest vm_accesssor_test
Bug: 358427181

Change-Id: Ib48e3ed61311c2eac0174f8315c2af4b59b7f6e4