Revert "ferrochrome-tests: Take screenshots while waiting for bo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/229e1023-8328-49a7-9052-dacd3b10ef24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic2426e08d96d15a07a7deed3954c0060208fd4cc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com